Split topics models into two files
