+++ src/web/modules/topics/models/__init__.py
@@ -1,3 +1,4 @@
-from .main import *
 from modules.topics.entrance.levels import *
 from modules.topics.entrance.steps import *
+from .checking import *
+from .questionnaire import *
